requirements.txt
panda3d
panda3d-gltf
panda3d-blend2bam
